refactor: move FilterConfig and application logic to glitch-context
- Move FilterConfig enum from CLI to glitch-context library
- Add add_from_config method to GlitchContext for easier filter orchestration
- Simplify main.rs and cli.rs by leveraging the new library-based configuration
